move attributes filter into qraphql-products-filter
