+++ saleor/graphql/product/filters.py
@@ -1,14 +1,15 @@
 from collections import defaultdict
-from typing import Dict, List, Optional
+from typing import Dict, Iterable, List, Optional
 
 import django_filters
 import graphene
-from django.db.models import F, Q, Subquery, Sum
+from django.db.models import Exists, F, OuterRef, Q, Subquery, Sum
 from django.db.models.functions import Coalesce
 from graphene_django.filter import GlobalIDFilter, GlobalIDMultipleChoiceFilter
 
-from ...product.filters import filter_products_by_attributes_values
 from ...product.models import (
+    AssignedProductAttribute,
+    AssignedVariantAttribute,
     Attribute,
     Category,
     Collection,
@@ -74,6 +75,31 @@ def _clean_product_attributes_filter_input(
         queries[attr_pk] += attr_val_pk
 
     return queries
+
+
+T_PRODUCT_FILTER_QUERIES = Dict[int, Iterable[int]]
+
+
+def filter_products_by_attributes_values(qs, queries: T_PRODUCT_FILTER_QUERIES):
+    filters = [
+        Q(
+            Exists(
+                AssignedProductAttribute.objects.filter(
+                    product__id=OuterRef("pk"), values__pk__in=values
+                )
+            )
+        )
+        | Q(
+            Exists(
+                AssignedVariantAttribute.objects.filter(
+                    variant__product__id=OuterRef("pk"), values__pk__in=values,
+                )
+            )
+        )
+        for values in queries.values()
+    ]
+
+    return qs.filter(*filters)
 
 
 def filter_products_by_attributes(qs, filter_value):
