+++ saleor/product/tests/test_product.py
@@ -7,8 +7,8 @@ from freezegun import freeze_time
 from prices import Money
 
 from ...account import events as account_events
+from ...graphql.product.filters import filter_products_by_attributes_values
 from .. import models
-from ..filters import filter_products_by_attributes_values
 from ..models import DigitalContentUrl
 from ..thumbnails import create_product_thumbnails
 from ..utils.attributes import associate_attribute_values_to_instance
